Manage memory for properties
The memory management attribute has four values: strong, weak, copy, unsafe_unretained/assign
For non-objects, we use assign. unsafe_unretained is like assign but for objects.
unsafe_unretain implies that user is managing the memory.
weak will cause the reference to be automatically set to nil when the object is deallocated
copy is used for instances when the object has a mutable sub-class. Prevents a mutable instance from being passed in and then being modified outside the object
Always declare a value for memory management since the default changes overtime.