Learnt about navigating with a UINavigationController
Updated Homepwner to use a UINavigationController as its root view controller.
Set the root view controller of the navigation controller to the TableViewController
Handled the didSelectRowAtIndex selector so that it pushed a detail view controller
onto the navigation controller's view stack.
Note it is a view stack, so only the top of the stack is displayed.