Learnt about passing data between view controller.
This implementation is simply a property of the view controller and we use the setter
when we create it. In this case , it would have been better to use an initializer that
takes the object to be viewed in detail.
We use the date to populate the fields when the message viewWillAppear is received.