Learnt about modal view controller and blocks
Updated code so that new items are added using a modal view
that allows it properties to be edited. Added support for ipda specific
model view styles and supporting how to redisplay the table view with blocks
since some modal styles on the ipad does not cause the table view to disappear.